+++ pytorch_lightning/trainer/connectors/profiler_connector.py
@@ -32,24 +32,16 @@ class ProfilerConnector:
     def __init__(self, trainer):
         self.trainer = trainer
 
-    def on_trainer_init(self, profiler: Union[BaseProfiler, bool, str]):
+    def on_trainer_init(self, profiler: Union[BaseProfiler, str]):
 
-        if profiler and not isinstance(profiler, (bool, str, BaseProfiler)):
+        if profiler and not isinstance(profiler, (str, BaseProfiler)):
             # TODO: Update exception on removal of bool
             raise MisconfigurationException(
                 "Only None, bool, str and subclasses of `BaseProfiler`"
                 " are valid values for `Trainer`'s `profiler` parameter."
                 f" Received {profiler} which is of type {type(profiler)}."
             )
-
-        if isinstance(profiler, bool):
-            rank_zero_warn(
-                "Passing a bool value as a `profiler` argument to `Trainer` is deprecated"
-                " and will be removed in v1.3. Use str ('simple' or 'advanced') instead.", DeprecationWarning
-            )
-            if profiler:
-                profiler = SimpleProfiler()
-        elif isinstance(profiler, str):
+        if isinstance(profiler, str):
             if profiler.lower() in PROFILERS:
                 profiler_class = PROFILERS[profiler.lower()]
                 profiler = profiler_class()